This change adds ondemand prices to be output alongside the relevant spot prices.
Most of the time when using spot prices you want to compare them to the ondemand cost for alerting and graphing.

Unfortunately aws does not provide an api for getting ondemand costs. In order for this to work I am pulling from https://raw.githubusercontent.com/powdahound/ec2instances.info/master/www/instances.json. This is the backend data from ec2instances.info. As this might not be 100% reliable I have only enabled this via a command line option which is off by default.

This change also adds logging and a verbose option.